Move header checking to metadata.
Add basic structure for artifacts writer.
